requirements.txt
# =============================================================================
# AgentSociety Platform — Unified Python Dependencies
# =============================================================================

# ── FastAPI Backend ──────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.7          # 0.0.6 is incompatible with starlette>=0.27

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
psycopg2-binary>=2.9.0
asyncpg==0.29.0
alembic==1.13.1

# ── Authentication ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1                    # used directly in auth_service.py (passlib removed — abandoned & broken with bcrypt>=4)

# ── Background Tasks ────────────────────────────────────────────────────────
celery[redis]==5.3.6

# ── Simulation Engine ───────────────────────────────────────────────────────
ray[default]==2.9.0
chromadb==0.5.23
tenacity==8.2.3
numpy>=1.24.0,<2.0.0             # <2.0 required by onnxruntime and ray 2.9
# onnxruntime is a transitive dependency of chromadb — do not pin directly
# to avoid Windows DLL conflicts; let chromadb resolve the correct version

# ── AI / LLM ────────────────────────────────────────────────────────────────
google-genai>=1.0.0

# ── Video Processing ────────────────────────────────────────────────────────
opencv-python>=4.9.0.80
Pillow>=10.2.0

# ── Messaging / Communication ───────────────────────────────────────────────
paho-mqtt>=1.6.1,<2.0.0         # v2 has breaking API changes; project uses v1 API
redis==5.0.1

# ── HTTP / Networking ───────────────────────────────────────────────────────
httpx>=0.26.0
aiofiles==23.2.1

# ── Validation & Config ─────────────────────────────────────────────────────
pydantic[email]>=2.9.0           # [email] pulls in email-validator for EmailStr
pydantic-settings>=2.1.0
python-dotenv==1.0.0
email-validator>=2.0.0           # required by pydantic EmailStr used in schemas.py

# ── AWS S3 (optional) ───────────────────────────────────────────────────────
boto3>=1.34.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.3
